# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Evatrack API",
        "description": "API Documentation for Evatrack Evacuation Management System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/centers/{centerId}/units": {
            "get": {
                "tags": [
                    "Accommodation Units"
                ],
                "summary": "Get all accommodation units for a center",
                "operationId": "05832c465318bd1780cafb5a0ac5985d",
                "parameters": [
                    {
                        "name": "centerId",
                        "in": "path",
                        "description": "Center ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Center not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Accommodation Units"
                ],
                "summary": "Create accommodation unit for center",
                "operationId": "72fedc1f204b35c726c5880b00b3c903",
                "parameters": [
                    {
                        "name": "centerId",
                        "in": "path",
                        "description": "Center ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type_id",
                                    "max_capacity"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "type_id": {
                                        "type": "integer"
                                    },
                                    "max_capacity": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Capacity limit exceeded"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/accommodation-types": {
            "get": {
                "tags": [
                    "Accommodation Units"
                ],
                "summary": "Get all accommodation types",
                "operationId": "1cc9e6ffded724c626a3f99545ded88d",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/centers/{centerId}/units/{unitId}": {
            "delete": {
                "tags": [
                    "Accommodation Units"
                ],
                "summary": "Delete accommodation unit",
                "operationId": "49a6f8eb2efc8cf2a47d892834141648",
                "parameters": [
                    {
                        "name": "centerId",
                        "in": "path",
                        "description": "Center ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "unitId",
                        "in": "path",
                        "description": "Unit ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted successfully"
                    },
                    "400": {
                        "description": "Unit has current occupants"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Unit not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Accommodation Units"
                ],
                "summary": "Update accommodation unit details",
                "operationId": "2ef59d18e54f752807bad35cc3c8ea9a",
                "parameters": [
                    {
                        "name": "centerId",
                        "in": "path",
                        "description": "Center ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "unitId",
                        "in": "path",
                        "description": "Unit ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "type_id": {
                                        "type": "integer"
                                    },
                                    "max_capacity": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Unit or Center not found"
                    },
                    "422": {
                        "description": "Capacity limit exceeded"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/barangays": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "List barangays",
                "description": "Get all barangays, optionally filtered by city.",
                "operationId": "42c1a86f782ef5aaf0693648aea3a959",
                "parameters": [
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Filter by city name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/barangays/{id}/sitios": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "List sitios under a barangay",
                "operationId": "1b9ee82ca4ec4cad789d523c7d25be51",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Barangay ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/sitios/{id}/puroks": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "List puroks under a sitio",
                "operationId": "12e7a2ad888316e34365b5ebb89bf787",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Sitio ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}